refactor: make EngineLines independent of AnalysisController
